[CHANGE] moved Guild Commands to `Guild` Cog
